Change rendering of Notes: no nesting
Changes how Notes are rendered. Rather than dislay notes as a nested
hierarchy annotated by noteType under a top-level 'Contents' label,
notes are now grouped by noteType, where each noteType is rendered at
the root level as if it is a first-order property of the bib.

This also changes the label given to Notes from 'Contents' to 'Notes'
(although with this change, it's more of an internal handle since the
noteType is now displayed in it's place). A 'Contents' field is added,
mapped to 'tableOfContents', which has been separately mapped. (I think
there may have been confusion between Notes and Contents because Notes
used to include data that now appears exclusively as tableOfContents.)

Finally, worth mentioning that all notes are now indexed as blanknodes,
so I've retired the special handling of notes not indexed as objects.

https://jira.nypl.org/browse/SCC-390