Change qa eb SSLCertificateId to correct value
This commit overwrites the SSLCertificateId config in .ebextensions with
a cert arn relevant for `qa` and `master` (production). This overrides
the SSLCertificateId merged from `development` (which is only relevant
for development). This is step 3 of the three step process to establish
permanently divergent eb config in development and qa/production, which
should allow us to continue to use our developent>qa>master deploy
strategy while also preserving necessary eb config in the separate
deployment branches. That three step process was started, and is
documented, here:

https://github.com/NYPL-discovery/discovery-front-end/pull/969

https://jira.nypl.org/browse/SCC-834