Fix ResultsCount to update when new searches are made and next page is pressed.
Previously, when a two searches were executed one after another and
neither returned results, the search key words would not update. So, for
example, searching for 'AAAAAAAAA' and 'BBBBBBBB' in succession would
display 'no results for AAAAAAAAA' both times. Also, pressing next or
previous page, one would always have displayed "displaying 1-50 of ...".
Both these errors were caused by ResultsCount having a
shouldComponentUpdate that prevented the component from updating if the
count did not change. This has been changed so that the component will
update once isLoading is false. This also required moving several of the
Action calls in the Search component into an ajax call so they would be
asynchronous. Otherwise, these calls would occur before isLoading was
set to true, generating additional renders with false combinations of
searchKeywords and count.

The tests 'Correctly updates on new Search' has been added to check the
shouldComponentUpdate method of the ResultsCount component. In the
Search tests, the tests 'should not update the searchKeywords before it
submits the request' and 'should update the searchKeywords after it
submits the request' have been added to enforce that changes to
searchKeywords in state occur asynchronously. Finally a small error in
the Search tests file has been corrected by moving the mock variable to
a lower scope.