Refactor redirecting for ineligible patrons
This commit breaks out refactoring of ineligible patrons to a separate
method called conditionallyRedirect, which returns a promise. This
allows the HoldRequest tests to be written in a more sensible manner,
which we have done (i.e. we directly manage the async order instead of
using .catch as a workaround)