Fix inability to directly clear creatorLiteral
Fix bug where creatorLiteral was missing from appConfig.defaultFilters,
leading to the SelectedFilters component to fail to acknowledge it as a
valid filter, leading to there being no direct way to clear creator
filters from search results listings. This fix ensures that, when a
creatorLiteral filter is active, there will be a button to remove the
filter on search results pages.

https://jira.nypl.org/browse/SCC-2034