Update data-api-client for on-site-edd
Updates nypl-data-api-client to ^v1.0.1 so that we can pass request
headers. This entailed changing a few api response handlers because the
NyplDataApiClient prior to v1.0.0 stripped the top-level `data` property
out of any response it returned. (See
https://github.com/NYPL-discovery/node-nypl-data-api-client/blob/master/CHANGELOG.md#breaking-changes)

Also enables Request button for items that are `requestable` regardless
of whether they are on-site or off.