SCC-2228 Add "Journal Title" as a search option
This adds Journal Title as a search option in the main SCC search bar. This is done by adding in an additional filter on the `issuance` field when the option is selected. This restricts the results to `urn:biblvel:s` which corresponds to `serial` records. The main query is still executed on the `title` fields.

To achieve this, a small check is inserted for the search type and, if found, the `selectedFilters` field is extended with the serial value. This overrides any other values set on the `issuance` field, which I believe gives the intended functionality. (e.g. selecting Journal Title should override other format filters to avoid any unexpected results)

An additional test has been added to verify that the proper query URL is generated for the API when this option is selected.