Fix logger to not swallow runtime errors in test
Our use of winston used to specify that both transports
"handleExceptions" unconditionally. This causes winston to attempt
to catch uncaught exceptions (like missing modules) and log them
through the transport (i.e. format them in the nypl json log
format). This seems/seemed a reasonable feature to enable at one
point because it ensures that uncaught exceptions are rendered as
json and thus queryable in CloudWatch via $.level. (If we do not
have winston catch excptions, uncaught exceptions will be logged
to stdout and CW in plain text, which means they may not be picked up by
metrics listening on $.level)

This changes the primary winston transport to not `handleExceptions`
when running the test suite so that missing modules and other uncaught
exceptions are dumped to stdout and mocha's nonzero exit code isn't
swallowed.