Centralize check for NYPL bnumbers
Based on PR feedback, adds a new utils.isNyplBnumber method for
determining whether or not a given bnumber is NYPL or not. Used when
determining whether or not to fetch annotated-marc for a bib and whether
or not to show the Full Details tab. (We should do neither for partner
bibs.)