Fix: use identifier number query in aggregations
Fixes minor issue: Previously, when performing a search by an
"identifier number" (e.g. oclc, lccn), the app made two DiscoveryAPI
calls for 1) the results and 2) the aggregations. But it wasn't passing
the identifier number to the aggregations query, essentially performing
an empty aggregations query, which is very slow. This artificiallyu
slowed down identifier lookups.

https://jira.nypl.org/browse/SCC-2876