Refactor Variable Instantiation for ShortenItem
The variable `shortenItems` can be simplified by checking if the pathname on the location object contains the string "all". If "all" is not found it will return -1, otherwise it exists. By comparing the result of indexOf to -1 we can set the value of shortenItem to the inverse of the comparison.

Ex: IndexOf !==1
 - If it does not equal -1 set to true, meaning found and must be shortened.
 - if it does equal -1 set to false, meaning not found and must be shortened.