Remove Bib check and Re-insatiation
This is not needed at all. Since bib must exist due to the check on line 72 there is no need to recheck if it exists.

Rechecking and setting to empty object does not do anything. Line 72 already checks it bib exists. It it didn't, it would redirect the patron to a 404.