Separate Complex Content Maping For Component Separation
The intention here is to reduce the size and complexity of the Big Page. This initial component is verbose with what it's trying to control. Rather than requiring the top level page to control lower components, instead let it pass down to those components any configuration properties they need to do their work.