Temporarily Set Common ENV Settings Each Build Mode
This is to remove the conflicting 'process.env' environment configuration set up. The DefinePlugin was catching an overwrite of the env object in the production build.

The intention is to correct this issue and set it in a ticket for tech debt. This is a working solution which will be resolved after a review.