Utilize BackToSearchResults Component To Test For It's Link
The BibPage component was replaced with the BackToSearchResults component for running the Back To Search Results link test. The original test was shallowly rendering the Bibpage, which worked fine with the original structure. However, the search results link was extracted out of the Bibpage and constructed into it's own component. The shallow method prevented the Link from being rendered and caused a "not found" error during test runs. By replacing the BibPage with the actual component we are able to produce the desired result.