Do Not Display Aeon Request Link In ER if Table Contains Aeon Request Buttons (#1835)
Commits: 
  - Refactor Variable Instantiation for Shorten Item
  - Update Lint Extends Order
  - Note: Filtered Items is Set Twice: Constructor && Render
  - Utilities to Remove Aeon URLs From Electronic Resources
  - Pluck Aeon Links from Top Bib Details ER Prop
  - Create Test Cases To Validate Desired ER Links In Top Bib Details
  - Add Comment To Bib Test
  - Remove Superfluous Comments
  - Define The Structure The Passed Parameters to The Pluck Resources
  - Return False If NO URL To Check Aeon Links

These utilities are used to validate:
  - featuredAeonList: if the desired items list to display has a functional aeon link
  - isAeonLink: if the URL qualifies as Aeon
  - pluckAeonLinksFromResource: Remove all Aeon Links from List

The main function is pluckAeonLinksFromResource. It will remove only the ERS from Resources array if they contain valid Aeon Links and the list of items is will display the Aeon Request Button.

The variable `shortenItems` can be simplified by checking if the pathname on the location object contains the string "all". If "all" is not found it will return -1, otherwise it exists. By comparing the result of index of to -1 we can set the value of `shortenItem` to the inverse of the comparison.

Ex: Index of !==1
 - If it does not equal -1 set to true, meaning found and must be shortened.
 - if it does equal -1 set to false, meaning not found and must be shortened.