Rebuild Webpack Dev Server Instantiation Logic
The Webpack CLI and the Webpack Dev Server can find the dev server configuration objects from the webpack configuration. This simplifies our logic and can extract non production code out of our production server

This is an incomplete extraction due to the deployment build process.  A Jira ticket will be created to note this for future refactor.