Refactor Storage to Use Local Storage w/ Tests (#1850)
Using Session Storage tuned out to not be the best approach of saving the users EDD form data. Due to the header's authentication functionality, the session state would be dropped since the browser's origin was changed to the encore login component. To avoid the session from being dropped we moved all logic to local storage. This allows for persistence the EDD form origin. How the component works now is to check if a remount was initiated, fetching the state from LS and apply the values to their relevant input fields if the items ID matches and the initialization time is within tolerant limits. 

Commits
 - Remove Using Storage on Component Constructure
 - Remove Session Storage In Favor Of Local Storage
 - Add Date Initialization For Local Storage
    - Assuming the user has been dormant for a while, track for how long. Remove storage on render if longer than 30 mins.
    - This has two affects.
        1. Prevent longer (old) standing queries from adding content to new queries.
        2. Prevents confusion on long standing forms. Keeping inputs clear.

 - Set Local Inputs If Within Time Context && Availability
 - Remove Exports of Helper Functions
 - Add Item Id to ED Form Prop
 - Use Local Storage in Testing for ED Form
 - Move Helper Functions to Helper JS File
 - Add Test to Support Local Storage on Rerender
 - Ensure Local Storage Is Available on Test Dom
 - Remove Component will Unmount from Local Storage Option
 - Up Time from 30 Mins to 2 Hours on Init Check
 - Update Failing Test on Travis
 - Add Support for Checking Storage on New Item Tests