Check User Logged In After Aeon Url Redirect Check
Per the conversation with @nonword (Paul), @danamansana (Daniel) and @seanredmond (Sean) the Discovery Web Server does not need to require a user to be logged in prior to checking if the request for the hold request is an special collections item.  We will redirect the user to login after the above check.