Move NewReuestHold Test Down A Dir to Unit
It seems travis was not picking up the glob patern in the test script and only running the test in the server file. To avoid this action, the file was moved down a directory.

It would be nice to separate the tests between server and client, however this will need to be investigated if we are using travis to test becuase of the above issue.