Rebuild Webpack Dev Server (#1857)
During configuration of the Node v14 bump the webpack development environment was changed. Unfortunately, the configuration is now a bit broken, but this PR corrects some of the issues. The changes created WebSocket issues, causing multiple errors inside the browsers log and Hot Module Replacement became useless. While the HMR is still not working the WebSocket issue has been resolved because we are now not using HMR :)