Update Syltes For All Three Buttons
This commit addresses the third button which was omitted in the original refactor.

Additionally, components for each request button were created inorder to separate internal ephemeral state each button may or may not have.

In order to accomodate the third "Make Appointment" Special Collections button, not only were components created to address each, but also the styles had to be updated. The updated styles did however break the breakpoint styles when rendering in the search results page.