Build Second Row for Table Items  (#1862)
Added support for the updated Request Buttons.

We now have three buttons opposed to one which a patron can select and be directed to the appropriate page. The three buttons are:

1. EDDButton
2. AeonButton
3. ReCAPButton
These buttons are a WIP in that some logic can be curated for DRY patterns and the logic for when a button is displayed is still being determined. However, for styling purposes and they're ready to do.

Some work to still be performed:

1. Button display logic is still being fine-tuned. Talk to @seanredmond.
2. EDD request page still needs to be cleared up.
3. The Button label links are still missing a defined href path.
4. Additional label text and options need to be cleared up.