Add Format All NPM Script To Use Prettier
This was left out for by accident. It's intention is for the developer to run the formatter.  There is a "." at the end of the format script. Though it may make more sense to remove this and allow the dev to inject the file path manually when running the formatter. We'd need to rely on the dev knowing this is an option.