Remove Lint Errors && Add Lint Script
This removes all eslint "Errors" from our project. Regarding reference to the "q" property of a query, we are forced to keep this even though it falls under the id-length rule. "q" was added as an exception.

Additionally, a dev can quickly check all lint errors (excluding warnings with the 'quiet' argument) at anytime with npm run lint.