Run File Save (Prettier) On Relevant Files (#1846)
Commits:
  - Run File Save (Prettier) On Relavant Files
  - Pretty Up Supported Files Which Relate to WIP
  - Run Prettier On Entire Project
  - Add Format All NPM Script To Use Prettier
  - Correct Failing Test
  - Remove Lint Errors && Add Lint Script

This removes all eslint "Errors" from our project. Regarding reference to the "q" property of a query, we are forced to keep this even though it falls under the id-length rule. "q" was added as an exception.

Additionally, a dev can quickly check all lint errors (excluding warnings with the 'quiet' argument) at any time with npm run lint.