Prevent Default Only If Non Aeon URL For Request Buttons
By preventing the default logic of the Request Buttons if the button is an Aeon, specRequestable link, we will effectivley be prepending the aeon link into our base urls context with router.context.push.  The button is an a tag under the hood, and we want it to direct the patron to the appropriate site when clicked.