Use Underscores Flatten Method
There is an issue with the servers Nodejs Version and the browsers Ecmascripts version. Since we are using Nodejs v10 on the server we cannot take advantage of some methods such as "flat" on the array. Nodejs v10 does not include this method. So, to work around this we are importing "flatten" from underscore.