Build && Incorporate Identifier Field into Definition Field Component (#1888)
For a cleaner approach to processing Identifiers the identifiers were incorporated into the Definition Field.

The processing of the field is first checked against valid identifier items. When a field is not valid it skips adding the label to the definition list. If valid, the DefinitionField component will process the item and apply the IdentifierField comp for render.