Set Direction for Parallel Fields (#1894)
When a text string is passed to Parallel Field determine the direction of the field to be displayed. If the strings character Unicode value is u200F we will display the field direction as rtl, otherwise ltr. The direction argument is placed on the html blocked tag. EX: <span dir='rtl' >{text}</span>