Construct Headings Refactor By Bool Filtering Array
No need to return a single object. We can instead return an array of the possible construct but reduce the indexes by any undefined or null values set when there isn't a Parent object on the Headings construct. This is acheived by checking for parent in the assiciated index of the returned array. If there isn't the value will be undefiend. The returned array will then be a single item array after the undefineds are filtered out.