SCC-2915: Component Updates (#1897)
* Rename Origin to Value for Clarity

* Rename Variable For Valid Identifer Function

* Remove Use Memo Reference in Bib Provider

The memo was not usefull in this situation. A TODO comment was left as for reference for another days resolution. It was resolved and unnecessarily committed to the file.

* Remove Note

A note was left in BibPage file future reference. This was removed. It asked why the client is being asked to checkForMoreItems on the Bib. This is unresolved.

* Integrate BibDetails Functional With Original BibDetails

Inorder to retain the the diff between the files it was asked to integrate the two files.

* Update Tabbed Tests to Pass

NOTE: The Component "Tabbed" is not used within this application though the tests are still active.

* Remove Use Memo From BibDetails

The use of useMemo here was not very valuable and made no impact on the performace of the application. It was removed.

* Add Array Utility Functions To Replace Underscore

* Remove TODO Reference && useMemo Implementation

The use of useMemo was unnecessary and therefor removed.

* Remove TODO Reference

* Ensure LinkableBibField Can Process Values w/ "@Id"

* Remove Commented TODOs

* Add Reference to Jira Tickets

* Add TODO Jira References

* Remove TODO

* Remove Property "Additional" from DefinitionField

The value was not needed here as no field used by the BibDetails is actually additional.

* Replace BibHoldings w/ Library Holdings

Moved the section element from BibHoldings into LibraryHoldings and referenced the Jira Ticket

* Remove Commented Note

* Update Failing Tests For BibDetails

* Ooops: Added the Reducer Accumulator Back In

* Removed Unnecessary Tests For Discussion

* Process Owning Institutions In Bib Details

* Replace Underscore IsEmpty Util w/ Cust Util

* Add CoreJS to Base Server Index File

This is so the server can utilize the core-js package and note crash while building the SSR bundle due to the Array method `flat`. We are currently running NodeJS 10.x which does not come with `Array.flat()`. When building, the server was crashing.

* Destructure Subject Heading Props

* Remove Generate LI Component From SubjectHeadings

* Use Fragments in Place of Div

This is a stylistic choice but to be consistent with the DefinitionField's structure the DIVs should not be available.

* Construct Headings Refactor By Bool Filtering Array

No need to return a single object. We can instead return an array of the possible construct but reduce the indexes by any undefined or null values set when there isn't a Parent object on the Headings construct. This is acheived by checking for parent in the assiciated index of the returned array. If there isn't the value will be undefiend. The returned array will then be a single item array after the undefineds are filtered out.

* Ues a Func to Construct the Value

This helps simplify the code as we can put most, not all, the logic around what the value will be inside an enclosed execution context.

* Check For Subject Literal && Compress the Headings

* Set a JSDoc Desription of the ValidIdentifier Function

* Use If Block To Wrap Subject Headings Component

* Move Normalization of Subject Literal To Field Map

Also, palce the function to normalize the literal inside the utils file.

* Build Linkable Fields For Subject Literals

Due to the complexity of Subject Literals and the variance of how Linkable Field works, a new property was added to supplement the filter query. FilterPath was added to LinkableField and the path had to be rebuilt inorder to accommodate the literals filter query.

* Add Descriptive Comments For Literals

* Ensure Tests Use Current Core JS

* Removeing Tests Which Are Not Relevant Anymore