Extract Parallels In Bib Details && Use Directional
Complications arose when using the Parallels Field Component inside the LinkableFields Component. The concern was how a list of values was being passed to LinkableField. This list then insterted multiple values inside an Anchor Tag which has a class on display: inline-block. All elements inside the Anchor Tag would then be displayed inline. This caused major issues with some fields.

Additionally, the manner in which the ParallelsFields component extracted and processed parrallels was confusing. For it's reusability, it became unclear as to which properties were required and which were not. It is clearly an Anti Pattern.

The solution is to extract the parallels within the BidDetails component and flatten the results inside DefinitionField. The individual values would then be passed to the DirectionalText Component in place of the ParallelsField Component.