Fix the requestability check in hold-request page
This changes the requestability check on the hold-request page from
checking `item.physRequestable` (formerly `item.available`) to checking
`item.eddRequestable` and `item.physRequestable`.

The `.physRequestable` property can not be checked alone because the
page currently presents both EDD and Phys delivery options. Separate
Request Buttons rollout will include removing the EDD option from this
page. Until that happens, we need to check for both `.physRequestable`
and `.eddRequestable` because some bibs (e.g. all on-site materials
available for EDD, like b16323928) are only EDD
requestable.

The reason to not revert to the original `.avilable` property is that it
was never quite the right choice for requestability anyway. It will
always be `true` when `physRequestable` and/or `eddRequestable` are
`true` because requestability requires an "available" status and
`.available` [is computed based on item
status](https://github.com/NYPL/discovery-front-end/blob/4b87da98c505093cb7a4687cd79a0241f6ed7811/src/app/utils/item.js#L216)
. But it seems likely that an item may not be `.eddRequestable` or
`.physRequestable` even though it's .available`.

So this PR changes the check to use the two `..Requestable`
properties relevant for the page in it's current manifestation: A page
presenting EDD and Phys delivery options.